changed hashcode and equals for @Entity classes suggestions made by JPA Buddy. Also some tidying up.
https://www.jpa-buddy.com/blog/best-practices-and-common-pitfalls/

+++ datavault-common/src/main/java/org/datavaultplatform/common/model/dao/custom/BaseCustomDAOImpl.java
@@ -28,7 +28,7 @@ public abstract class BaseCustomDAOImpl implements BaseCustomDAO {
 
   protected long getCount(TypedQuery<Long> typedQuery) {
     Long result = typedQuery.getSingleResult();
-    return result == null ? 0 : result.longValue();
+    return result == null ? 0 : result;
   }
 
   protected <V> List<V> getResults(CriteriaQuery<V> cq) {
